pyproject.toml
[project]
name = "ki-geodaten"
version = "0.1.0"
description = "Text-to-Polygon Pipeline — DOP20 + SAM 3.1"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.7",
    "torchvision",
    "numpy>=1.26,<2",
    "rasterio>=1.4,<1.5",
    "geopandas>=1.0",
    "shapely>=2.0",
    "pyproj>=3.6",
    "fastapi",
    "uvicorn[standard]",
    "jinja2",
    "pydantic>=2",
    "pydantic-settings",
    "requests",
    "urllib3>=2",
    "pillow",
    "fiona",
    "transformers>=5.6",
    "accelerate",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-mock", "responses", "httpx"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["ki_geodaten*"]
